Include abctree.py and treeslice.py from bintrees into this project
The bintrees project states that it is deprecated on its github page. Since
accumulation_tree is closely intertwined with this software, it makes most
sense to just fork those parts we depend on. This also makes packaging easier.

These files are MIT-licensed, which is compatible with including these into a
GPL project.